Add extensions command
New extensions command to report on extensions specifically.

Supports splitting of extensions having multiple parts.

It is also a use-case for the refactored path iteration. The Paths object
is extracted based on commons with the report command.

The report itself has become an interface and the generic reports as well
as the new extensions report and their collaborators are now covered.

Refs:

- Command: extensions

- Command: report