Prune directories in recursive iteration via Matcher
An optional Matcher can be provided to the INodeIterFactory which will
filter all created recursive directory iterators matching the filename
of the directory by that Matcher.

For a default implementation the PatternMatcher has been created which
support common file-name patterns with * to match any character but "/",
? to match a single character but "/". A pattern prefixed with ! will
remove that pattern from the Matcher, if "!" at the beginning is part of
the pattern, prefix it with "\" (backslash).

Multiple patterns can be separated with semicolon (";") to have the semi-
colon within a pattern, prefix it with "\" (backslash).

Only exact patterns that are already part of the Matcher can be removed
with "!".